feat: added a request for deleting a tag in a collection

+++ server/collection-manager-service/src/main/java/ru/vesuvian/collection/dto/get/TagGetDto.java
@@ -23,7 +23,7 @@ public class TagGetDto {
 
     @Schema(description = "Tag ID")
     @JsonProperty("id")
-    private Long tagId;
+    private Integer tagId;
 
     @Schema(description = "Tag name")
     @JsonProperty("name")
